+++ src/main/java/org/deeplearning4j/gym/Client.java
@@ -1,11 +1,10 @@
-package org.deeplearning4j.rl4j;
+package org.deeplearning4j.gym;
 
 
 import com.mashape.unirest.http.JsonNode;
 import lombok.Value;
-import org.deeplearning4j.rl4j.gym.space.GymObservationSpace;
-import org.deeplearning4j.rl4j.space.ActionSpace;
-import org.deeplearning4j.rl4j.space.ObservationSpace;
+import org.deeplearning4j.gym.space.GymObservationSpace;
+import org.deeplearning4j.gym.space.ActionSpace;
 import org.json.JSONObject;
 
 import java.util.Set;
