+++ src/main/java/org/deeplearning4j/gym/space/Box.java
@@ -1,6 +1,6 @@
-package org.deeplearning4j.rl4j.gym.space;
+package org.deeplearning4j.gym.space;
 
-import org.deeplearning4j.rl4j.space.Encodable;
+import org.deeplearning4j.gym.space.Encodable;
 import org.json.JSONArray;
 
 /**
