+++ src/main/java/org/deeplearning4j/gym/space/DiscreteSpace.java
@@ -1,4 +1,4 @@
-package org.deeplearning4j.rl4j.space;
+package org.deeplearning4j.gym.space;
 
 import lombok.Getter;
